Split tag routes out of the highlights router
Eight endpoints — the tag CRUD, both tag group routes and both tag-highlight
association routes — were defined in the reading highlights router, so they
inherited its tags=["highlights"]. Tags have no dedicated router, and therefore
no OpenAPI tag of their own.

That grouping reaches the frontend. orval generates in tags-split mode, which
splits modules by OpenAPI tag, so every tag operation landed in
generated/highlights/highlights.ts and any caller working with tags had to
import from the highlights module. The eight now generate into their own.

This is a router split, not a domain move. Tag stays under reading/: both Tag
and TagGroup carry a book_id and have no lifetime independent of a Book, so
where the aggregate ultimately belongs is a separate question that this does not
prejudge. An OpenAPI tag is a client-grouping hint, and grouping the endpoints
does not require answering it.

Paths and handler names are carried over unchanged, so FastAPI derives identical
operationIds and generated clients see only the tag change. That is also why the
two tag group routes still sit under /highlights, which now reads oddly — moving
them to /tag-groups would rename their operationIds and break callers, so it is
left as a deliberate follow-up rather than smuggled in here.

Verified by diffing the OpenAPI operation table before and after: no routes
added or removed, no operationId changed, exactly eight tag values flipped from
"highlights" to "tags". 517 tests pass, including tests/test_tag_groups.py,
which exercises these routes by URL.
